fix: restructured and fixes error breaking the app
