+++ package.json
@@ -27,8 +27,8 @@
     "stream": "^0.0.2"
   },
   "devDependencies": {
-    "@azure/openai": "^1.0.0-beta.11",
     "@anthropic-ai/sdk": "^0.18.0",
+    "@azure/openai": "^1.0.0-beta.11",
     "@mistralai/mistralai": "^0.1.3",
     "chai": "^5.0.3",
     "cohere-ai": "^7.7.3",
